warn user when cb fails to copy
I did some testing and it seems like if you had copied a larger file prior to taking screenshots, cb will not have space left to copy the screenshot. This is weird because cb clear Clipboard before trying to copy the screenshot. This is rare, if this happens try doing `cb info` and rming the "Stored in" folder. ex: `rm -rf /run/user/1000/Clipboard/0/*`. I don't know any alternative to this. The screenshots are still saved though, which is less worse.